pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mini-webui"
version = "0.1.0"
description = "delivers a streamlined AI chat console for teams that need rapid iteration, reliable integrations, and production-ready guardrails"
authors = [
    {name = "mini-webui Team"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "openai>=1.3.7",
    "pydantic[all,email]>=2.5.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "tiktoken>=0.7.0",
]
requires-python = ">=3.9"

[project.scripts]
mini-webui = "mini_webui.main:main"

[tool.hatch.build.targets.wheel]
packages = ["backend/mini_webui"]
